register and unregister user for course throughout course, courserepository, courseservice and coursecontroller

+++ src/main/java/org/example/greenpointbackend/repository/CourseRepository.java
@@ -8,7 +8,8 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
     @Repository
-    public interface CourseRepository extends JpaRepository<Course, Long> {
+    public interface CourseRepository extends JpaRepository<Course, Integer> {
         List<Course> findCoursesByJobTitle(JobTitle jobTitle);
+
     }
